Moving all DAO to common/dao Moving all test data to common-test
