+++ src/common-test/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "ccc-aws-common-test",
+  "version": "1.0.1",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC"
+}
