+++ src/common/dao/base.js
@@ -1,4 +1,4 @@
-const db = require('./db');
+const db = require('../db');
 
 class BaseDAO {
     constructor(configuration, defaultTableName) {
