+++ src/common/dao/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+    BaseDAO: require('./base'),
+    infractions: require('./infractions'),
+    devices: require('./devices'),
+    scoring: require('./scoring'),
+    siteQuality: require('./siteQuality')
+};
