+++ src/validate/package.json
@@ -17,11 +17,12 @@
   },
   "homepage": "https://github.com/Cyberspace-Citizenship-Coefficient/AWS-Lambdas#readme",
   "dependencies": {
-    "ccc-aws-lambda-common": "^1.0.15",
+    "ccc-aws-common": "^1.0.23",
     "puppeteer": "^10.1.0",
     "vandium": "^6.1.0"
   },
   "devDependencies": {
+    "ccc-aws-common-test": "^1.0.1",
     "chai": "^4.3.4",
     "grunt": "^1.4.1",
     "grunt-aws-lambda": "^0.13.0",
